show mobile no

+++ flaskpostgresproject/models.py
@@ -29,4 +29,4 @@ class User(db.Model):
  email = db.Column(db.String(120), unique=True, nullable=False)
  image_file = db.Column(db.String(20), nullable=False, default='default.jpg')
  password = db.Column(db.String(60), nullable=False)
- mobile_no = db.Column(db.Integer(11), nullable=False)
+ mobile_no = db.Column(db.Integer, nullable=False)
